+++ packages/cryptoplease/lib/app/screens/authenticated/routes.dart
@@ -11,13 +11,14 @@ import 'package:cryptoplease/features/app_lock/presentation/routes.dart';
 import 'package:cryptoplease/features/backup_phrase/presentation/routes.dart';
 import 'package:cryptoplease/features/legal/legal_document_view.dart';
 import 'package:cryptoplease/features/nft/presentation/nft_details_screen.dart';
-import 'package:cryptoplease/features/nft/presentation/nft_screen.dart';
 import 'package:cryptoplease/features/outgoing_transfer/presentation/outgoing_transfer_flow/outgoing_transfer_flow.dart';
 import 'package:cryptoplease/features/outgoing_transfer/presentation/send_flow/routes.dart';
 import 'package:cryptoplease/features/profile/edit_profile_screen.dart';
 import 'package:cryptoplease/features/profile/help_screen.dart';
 import 'package:cryptoplease/features/profile/menu_screen.dart';
 import 'package:cryptoplease/features/profile/profile_screen.dart';
+import 'package:cryptoplease/features/request_pay/presentation/request_pay_flow.dart';
+import 'package:cryptoplease/features/request_pay/presentation/routes.dart';
 import 'package:cryptoplease/features/swap_tokens/presentation/routes.dart';
 import 'package:cryptoplease/features/swap_tokens/presentation/swap_token_flow.dart';
 import 'package:cryptoplease/features/transactions/presentation/transaction_details_screen/transaction_details_screen.dart';
@@ -30,11 +31,14 @@ const authenticatedFlowRoutes = AutoRoute<void>(
       path: '',
       page: HomeTabsScreen,
       children: [
-        CustomRoute<void>(path: '', page: WalletScreen),
-        CustomRoute<void>(page: NftScreen),
+        CustomRoute<void>(path: '', page: MenuScreen),
+        CustomRoute<void>(
+          page: RequestPayFlowScreen,
+          children: requestPayRoutes,
+        ),
         CustomRoute<void>(page: ActivitiesScreen),
         CustomRoute<void>(page: SwapTokenFlowScreen, children: swapTokenRoutes),
-        CustomRoute<void>(page: MenuScreen),
+        CustomRoute<void>(page: WalletScreen),
         CustomRoute<Token>(page: TokenSelectorScreen),
       ],
     ),
@@ -47,7 +51,7 @@ const authenticatedFlowRoutes = AutoRoute<void>(
     ...receiveFlowRoutes,
     ...addFundsRoutes,
     appLockSetupFlowRoutes,
-    AutoRoute<void>(page: ProfileScreen),
+    AutoRoute<void>(page: ProfileScreen, fullscreenDialog: true),
     AutoRoute<void>(page: EditProfileScreen),
     AutoRoute<void>(page: TermsOfServiceScreen),
     AutoRoute<void>(page: PrivacyPolicyScreen),