+++ packages/cryptoplease/lib/features/request_pay/bl/request_pay_event.dart
@@ -1,0 +1,6 @@
+part of 'request_pay_bloc.dart';
+
+@freezed
+class RequestPayEvent with _$RequestPayEvent {
+  const factory RequestPayEvent.amountUpdated(Decimal amount) = _AmountUpdated;
+}